SAK-52677 LTI move token validation out of servlet
